+++ sys/arm/rockchip/files.rk32xx
@@ -27,3 +27,4 @@ dev/iicbus/pmic/act8846_regulator.c	standard
 dev/iicbus/pmic/fan53555.c		standard
 dev/iicbus/rtc/hym8563.c		standard
 dev/mmc/host/dwmmc_rockchip.c		optional	dwmmc
+dev/sdhci/sdhci_fdt_rockchip.c		optional	sdhci
